H2-M2/requirements.txt
flask
python-dotenv
ipykernel
torch
pillow
transformers
openai